requirements.txt
streamlit>=1.38.0
sqlalchemy>=2.0.0
# psycopg2-binary>=2.9.9  # uncomment if using PostgreSQL
bcrypt>=4.1.0
python-dotenv>=1.0.0
pydantic>=2.5.0
openai>=1.12.0
pydub>=0.25.1
Pillow>=10.0.0
requests>=2.31.0
stripe>=8.0.0
streamlit_extras
